refator: notification 이벤트, 비동기 중첩 제거
