+++ src/main/java/com/sprint/mission/discodeit/service/NotificationService.java
@@ -1,6 +1,7 @@
 package com.sprint.mission.discodeit.service;
 
 import com.sprint.mission.discodeit.dto.response.NotificationResponse;
+import com.sprint.mission.discodeit.event.NotificationEvent;
 import java.util.List;
 import java.util.UUID;
 
@@ -9,4 +10,6 @@ public interface NotificationService {
     List<NotificationResponse> getMyNotifications(UUID receiverId);
 
     void readNotification(UUID id, UUID receiverId);
+
+    void createNotification(NotificationEvent notificationEvent);
 }
